chore(wrangler): Removed refactoring of bulk get file.
