--- packages/wrangler/src/kv/cli/bulkGet.ts
@@ -1,140 +1,0 @@
-import { readConfig } from "../../config";
-import { createCommand } from "../../core/create-command";
-import { UserError } from "../../errors";
-import { logger } from "../../logger";
-import * as metrics from "../../metrics";
-import { parseJSON, readFileSync } from "../../parse";
-import { requireAuth } from "../../user";
-import { isLocal } from "../../utils/is-local";
-import { getKVBulkKeyValue } from "../fetchers/getKVBulkKeyValue";
-import { getKVNamespaceId, usingLocalNamespace } from "../helpers";
-import type { EventNames } from "../../metrics";
-
-export const kvBulkGetCommand = createCommand({
-	metadata: {
-		description: "Gets multiple key-value pairs from a namespace",
-		status: "open-beta",
-		owner: "Product: KV",
-	},
-
-	positionalArgs: ["filename"],
-	args: {
-		filename: {
-			describe: "The file containing the keys to get",
-			type: "string",
-			demandOption: true,
-		},
-		binding: {
-			type: "string",
-			requiresArg: true,
-			describe: "The name of the namespace to get from",
-		},
-		"namespace-id": {
-			type: "string",
-			requiresArg: true,
-			describe: "The id of the namespace to get from",
-		},
-		preview: {
-			type: "boolean",
-			describe: "Interact with a preview namespace",
-		},
-		local: {
-			type: "boolean",
-			describe: "Interact with local storage",
-		},
-		remote: {
-			type: "boolean",
-			describe: "Interact with remote storage",
-			conflicts: "local",
-		},
-		"persist-to": {
-			type: "string",
-			describe: "Directory for local persistence",
-		},
-	},
-
-	async handler({ filename, ...args }) {
-		const localMode = isLocal(args);
-		const config = readConfig(args);
-		const namespaceId = getKVNamespaceId(args, config);
-
-		const content = parseJSON(readFileSync(filename), filename) as (
-			| string
-			| { name: string }
-		)[];
-
-		if (!Array.isArray(content)) {
-			throw new UserError(
-				`Unexpected JSON input from "${filename}".\n` +
-					`Expected an array of strings but got:\n${content}`
-			);
-		}
-
-		const errors: string[] = [];
-
-		const keysToGet: string[] = [];
-		for (const [index, item] of content.entries()) {
-			const key = typeof item !== "string" ? item?.name : item;
-
-			if (typeof key !== "string") {
-				errors.push(
-					`The item at index ${index} is type: "${typeof item}" - ${JSON.stringify(
-						item
-					)}`
-				);
-				continue;
-			}
-			keysToGet.push(key);
-		}
-
-		if (errors.length > 0) {
-			throw new UserError(
-				`Unexpected JSON input from "${filename}".\n` +
-					`Expected an array of strings or objects with a "name" key.\n` +
-					errors.join("\n")
-			);
-		}
-
-		let metricEvent: EventNames;
-		if (localMode) {
-			const result = await usingLocalNamespace(
-				args.persistTo,
-				config,
-				namespaceId,
-				async (namespace) => {
-					const out = {} as { [key: string]: { value: string | null } };
-					for (const key of keysToGet) {
-						const value = await namespace.get(key, "text");
-
-						out[key as string] = {
-							value,
-						};
-					}
-					return out;
-				}
-			);
-
-			logger.log(JSON.stringify(result, null, 2));
-
-			metricEvent = "get kv key-values (bulk) (local)";
-		} else {
-			const accountId = await requireAuth(config);
-
-			logger.log(
-				JSON.stringify(
-					await getKVBulkKeyValue(accountId, namespaceId, keysToGet),
-					null,
-					2
-				)
-			);
-
-			metricEvent = "get kv key-values (bulk)";
-		}
-
-		metrics.sendMetricsEvent(metricEvent, {
-			sendMetrics: config.send_metrics,
-		});
-
-		logger.log("\nSuccess!");
-	},
-});
