+++ packages/wrangler/src/kv/helpers.ts
@@ -286,6 +286,32 @@ function logBulkProgress(
 	);
 }
 
+type BulkGetResponse = {
+	values: {
+		[key: string]: {
+			value: string | object | null;
+			metadata?: object;
+		};
+	};
+};
+
+export async function getKVBulkKeyValue(
+	accountId: string,
+	namespaceId: string,
+	keys: string[]
+) {
+	const requestPayload = { keys };
+
+	return await fetchResult<BulkGetResponse>(
+		`/accounts/${accountId}/storage/kv/namespaces/${namespaceId}/bulk/get`,
+		{
+			method: "POST",
+			body: JSON.stringify(requestPayload),
+			headers: { "Content-Type": "application/json" },
+		}
+	);
+}
+
 export async function putKVBulkKeyValue(
 	accountId: string,
 	namespaceId: string,
